Upgrade role to support py3
Remove get-pip.py as it was ancient and not compatible with py3.
Use upstream version every time instead.

Add defaults for ansible_ssh_user and ansible_python_interpreter

rework bootstrap.sh as install_python.sh to download/build off
version variables with defaults as:
 - python_version: 3.5
 - pypy_version: 6.0.0

Move the setting of the .bootstrapped filed into ansible and
have the name reflect the python and pypy version so that python
is reinstalled even a version is updated.

Move pypy install into /opt/python so it is installed 'globally'
rather than just for the core user.

Set min ansible version 2.5 (for py3 support).